// src/app/models/financial.ts
export type factorItem = {
  id?: number;
  code: string | null;
  name: string | null;
  countType: string | null;
  unitPrice: string | null;
  offPrice?: string | null;
  additionalPrice?: string | null;
  description?: string | null;
  count: string | null;
  width?: string | null;
  height?: string | null;
}